Move Chromium browser resolution helpers
