+++ src/Aspire.Hosting/BrowserLogs/BrowserLogsRunningSession.cs
@@ -5,7 +5,6 @@
 
 using System.Globalization;
 using System.Text;
-using System.Text.Json;
 using Microsoft.Extensions.Logging;
 
 namespace Aspire.Hosting;
@@ -327,254 +326,6 @@ internal sealed class BrowserLogsRunningSession : IBrowserLogsRunningSession
         }
     }
 
-    internal static bool IsGoogleChromeDefaultUserDataDirectory(string browser, string browserExecutable, string userDataDirectory)
-    {
-        if (GetBrowserKind(browser, browserExecutable) != BrowserKind.Chrome ||
-            MatchesBrowser(browser, browserExecutable, "chromium", "chromium-browser") ||
-            TryResolveBrowserUserDataDirectory(browser, browserExecutable) is not { } defaultUserDataDirectory)
-        {
-            return false;
-        }
-
-        var comparer = OperatingSystem.IsWindows() ? StringComparer.OrdinalIgnoreCase : StringComparer.Ordinal;
-        return comparer.Equals(NormalizePath(userDataDirectory), NormalizePath(defaultUserDataDirectory));
-
-        static string NormalizePath(string path) => Path.TrimEndingDirectorySeparator(Path.GetFullPath(path));
-    }
-
-    internal static string? TryResolveBrowserExecutable(string browser)
-    {
-        if (Path.IsPathRooted(browser) && File.Exists(browser))
-        {
-            return browser;
-        }
-
-        foreach (var candidate in GetBrowserCandidates(browser))
-        {
-            if (Path.IsPathRooted(candidate))
-            {
-                if (File.Exists(candidate))
-                {
-                    return candidate;
-                }
-            }
-            else if (PathLookupHelper.FindFullPathFromPath(candidate) is { } resolvedPath)
-            {
-                return resolvedPath;
-            }
-        }
-
-        return PathLookupHelper.FindFullPathFromPath(browser);
-    }
-
-    internal static string? TryResolveBrowserUserDataDirectory(string browser, string browserExecutable)
-    {
-        var browserKind = GetBrowserKind(browser, browserExecutable);
-        if (browserKind == BrowserKind.Unknown)
-        {
-            return null;
-        }
-
-        if (OperatingSystem.IsMacOS())
-        {
-            var home = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-            return browserKind switch
-            {
-                BrowserKind.Edge => Path.Combine(home, "Library", "Application Support", "Microsoft Edge"),
-                BrowserKind.Chrome => Path.Combine(home, "Library", "Application Support", "Google", "Chrome"),
-                _ => null
-            };
-        }
-
-        if (OperatingSystem.IsWindows())
-        {
-            var localAppData = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
-            return browserKind switch
-            {
-                BrowserKind.Edge => Path.Combine(localAppData, "Microsoft", "Edge", "User Data"),
-                BrowserKind.Chrome => Path.Combine(localAppData, "Google", "Chrome", "User Data"),
-                _ => null
-            };
-        }
-
-        var homeDirectory = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-        return browserKind switch
-        {
-            BrowserKind.Edge => Path.Combine(homeDirectory, ".config", "microsoft-edge"),
-            BrowserKind.Chrome => Path.Combine(
-                homeDirectory,
-                ".config",
-                MatchesBrowser(browser, browserExecutable, "chromium", "chromium-browser") ? "chromium" : "google-chrome"),
-            _ => null
-        };
-    }
-
-    internal static string ResolveBrowserProfileDirectory(string userDataDirectory, string profile)
-    {
-        ArgumentException.ThrowIfNullOrWhiteSpace(userDataDirectory);
-        ArgumentException.ThrowIfNullOrWhiteSpace(profile);
-
-        if (!Directory.Exists(userDataDirectory))
-        {
-            throw new InvalidOperationException($"Browser user data directory '{userDataDirectory}' was not found.");
-        }
-
-        if (TryResolveBrowserProfileDirectoryFromDirectoryEntries(userDataDirectory, profile) is { } directMatch)
-        {
-            return directMatch;
-        }
-
-        var localStatePath = Path.Combine(userDataDirectory, "Local State");
-        if (File.Exists(localStatePath))
-        {
-            try
-            {
-                using var localStateStream = File.OpenRead(localStatePath);
-                using var localStateDocument = JsonDocument.Parse(localStateStream);
-                if (TryResolveBrowserProfileDirectory(localStateDocument.RootElement, userDataDirectory, profile) is { } profileDirectory)
-                {
-                    return profileDirectory;
-                }
-            }
-            catch (IOException ex)
-            {
-                throw new InvalidOperationException(
-                    $"Unable to read Chromium profile metadata from '{localStatePath}' while resolving browser profile '{profile}'.",
-                    ex);
-            }
-            catch (UnauthorizedAccessException ex)
-            {
-                throw new InvalidOperationException(
-                    $"Unable to read Chromium profile metadata from '{localStatePath}' while resolving browser profile '{profile}'.",
-                    ex);
-            }
-            catch (JsonException ex)
-            {
-                throw new InvalidOperationException(
-                    $"Chromium profile metadata in '{localStatePath}' is invalid while resolving browser profile '{profile}'.",
-                    ex);
-            }
-        }
-
-        throw new InvalidOperationException(
-            $"Browser profile '{profile}' was not found under '{userDataDirectory}'. Specify the profile directory name (for example 'Default' or 'Profile 1') or a browser profile name from Chromium's profile metadata.");
-    }
-
-    internal static string? TryResolveBrowserProfileDirectory(JsonElement localStateRoot, string userDataDirectory, string profile)
-    {
-        ArgumentException.ThrowIfNullOrWhiteSpace(userDataDirectory);
-        ArgumentException.ThrowIfNullOrWhiteSpace(profile);
-
-        if (!localStateRoot.TryGetProperty("profile", out var profileElement) ||
-            !profileElement.TryGetProperty("info_cache", out var infoCacheElement) ||
-            infoCacheElement.ValueKind != JsonValueKind.Object)
-        {
-            return null;
-        }
-
-        string? match = null;
-
-        foreach (var profileEntry in infoCacheElement.EnumerateObject())
-        {
-            if (!Directory.Exists(Path.Combine(userDataDirectory, profileEntry.Name)) ||
-                !MatchesBrowserProfile(profileEntry, profile))
-            {
-                continue;
-            }
-
-            if (match is not null && !string.Equals(match, profileEntry.Name, StringComparison.Ordinal))
-            {
-                throw new InvalidOperationException(
-                    $"Browser profile '{profile}' matched multiple Chromium profiles under '{userDataDirectory}'. Specify the profile directory name instead.");
-            }
-
-            match = profileEntry.Name;
-        }
-
-        return match;
-    }
-
-    private static IEnumerable<string> GetBrowserCandidates(string browser)
-    {
-        if (OperatingSystem.IsMacOS())
-        {
-            return browser.ToLowerInvariant() switch
-            {
-                "msedge" or "edge" =>
-                [
-                    "/Applications/Microsoft Edge.app/Contents/MacOS/Microsoft Edge",
-                    "msedge"
-                ],
-                "chrome" or "google-chrome" =>
-                [
-                    "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome",
-                    "google-chrome",
-                    "chrome"
-                ],
-                _ => [browser]
-            };
-        }
-
-        if (OperatingSystem.IsWindows())
-        {
-            var programFiles = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles);
-            var programFilesX86 = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86);
-
-            return browser.ToLowerInvariant() switch
-            {
-                "msedge" or "edge" =>
-                [
-                    Path.Combine(programFilesX86, "Microsoft", "Edge", "Application", "msedge.exe"),
-                    Path.Combine(programFiles, "Microsoft", "Edge", "Application", "msedge.exe"),
-                    "msedge.exe"
-                ],
-                "chrome" or "google-chrome" =>
-                [
-                    Path.Combine(programFilesX86, "Google", "Chrome", "Application", "chrome.exe"),
-                    Path.Combine(programFiles, "Google", "Chrome", "Application", "chrome.exe"),
-                    "chrome.exe"
-                ],
-                _ => [browser]
-            };
-        }
-
-        return browser.ToLowerInvariant() switch
-        {
-            "msedge" or "edge" => ["microsoft-edge", "microsoft-edge-stable", "msedge"],
-            "chrome" or "google-chrome" => ["google-chrome", "google-chrome-stable", "chrome", "chromium-browser", "chromium"],
-            _ => [browser]
-        };
-    }
-
-    private static string? TryResolveBrowserProfileDirectoryFromDirectoryEntries(string userDataDirectory, string profile)
-    {
-        foreach (var directoryPath in Directory.EnumerateDirectories(userDataDirectory))
-        {
-            var directoryName = Path.GetFileName(directoryPath);
-            if (string.Equals(directoryName, profile, StringComparison.OrdinalIgnoreCase))
-            {
-                return directoryName;
-            }
-        }
-
-        return null;
-    }
-
-    private static BrowserKind GetBrowserKind(string browser, string browserExecutable)
-    {
-        if (MatchesBrowser(browser, browserExecutable, "msedge", "edge", "microsoft-edge"))
-        {
-            return BrowserKind.Edge;
-        }
-
-        if (MatchesBrowser(browser, browserExecutable, "chrome", "google-chrome", "chromium", "chromium-browser"))
-        {
-            return BrowserKind.Chrome;
-        }
-
-        return BrowserKind.Unknown;
-    }
-
     internal static string? TrySelectTrackedTargetId(IReadOnlyList<BrowserLogsTargetInfo>? targetInfos)
     {
         if (targetInfos is null)
@@ -600,38 +351,6 @@ internal sealed class BrowserLogsRunningSession : IBrowserLogsRunningSession
             targetInfo.Attached != true &&
             !string.IsNullOrWhiteSpace(targetInfo.TargetId))
             ?.TargetId;
-    }
-
-    private static bool MatchesBrowser(string browser, string browserExecutable, params string[] names)
-    {
-        var browserLower = browser.ToLowerInvariant();
-        var executableLower = browserExecutable.ToLowerInvariant();
-
-        foreach (var name in names)
-        {
-            if (browserLower == name ||
-                Path.GetFileNameWithoutExtension(browserLower) == name ||
-                executableLower.Contains(name, StringComparison.Ordinal))
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private static bool MatchesBrowserProfile(JsonProperty profileEntry, string profile)
-    {
-        return string.Equals(profileEntry.Name, profile, StringComparison.OrdinalIgnoreCase) ||
-            MatchesBrowserProfileProperty(profileEntry.Value, "name", profile) ||
-            MatchesBrowserProfileProperty(profileEntry.Value, "shortcut_name", profile);
-    }
-
-    private static bool MatchesBrowserProfileProperty(JsonElement profileElement, string propertyName, string profile)
-    {
-        return profileElement.TryGetProperty(propertyName, out var propertyElement) &&
-            propertyElement.ValueKind == JsonValueKind.String &&
-            string.Equals(propertyElement.GetString(), profile, StringComparison.OrdinalIgnoreCase);
     }
 
     internal static string BuildCommandLine(IReadOnlyList<string> arguments)
@@ -707,13 +426,6 @@ internal sealed class BrowserLogsRunningSession : IBrowserLogsRunningSession
     }
 
     private sealed record BrowserSessionResult(int? ExitCode, Exception? Error);
-
-    private enum BrowserKind
-    {
-        Unknown,
-        Edge,
-        Chrome
-    }
 
 }
