+++ src/Aspire.Hosting/BrowserLogs/ChromiumBrowserResolver.cs
@@ -1,0 +1,335 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Text.Json;
+
+namespace Aspire.Hosting;
+
+/// <summary>
+/// Resolves Chromium-based browser executables, user data directories, and profile directories.
+/// </summary>
+/// <remarks>
+/// This type translates the resolved browser-log configuration into local machine paths. Keep OS/browser probing here
+/// so <see cref="BrowserConfiguration"/> stays focused on configuration precedence and effective option values.
+/// </remarks>
+internal static class ChromiumBrowserResolver
+{
+    /// <summary>
+    /// Returns whether the requested path is branded Google Chrome's default user data directory.
+    /// </summary>
+    internal static bool IsGoogleChromeDefaultUserDataDirectory(string browser, string browserExecutable, string userDataDirectory)
+    {
+        // Google Chrome rejects remote debugging against its real default profile root. Chromium builds can still use the
+        // same resolver path, so exclude Chromium aliases before comparing with Chrome's default user data directory.
+        if (GetBrowserKind(browser, browserExecutable) != BrowserKind.Chrome ||
+            MatchesBrowser(browser, browserExecutable, "chromium", "chromium-browser") ||
+            TryResolveUserDataDirectory(browser, browserExecutable) is not { } defaultUserDataDirectory)
+        {
+            return false;
+        }
+
+        var comparer = OperatingSystem.IsWindows() ? StringComparer.OrdinalIgnoreCase : StringComparer.Ordinal;
+        return comparer.Equals(NormalizePath(userDataDirectory), NormalizePath(defaultUserDataDirectory));
+
+        static string NormalizePath(string path) => Path.TrimEndingDirectorySeparator(Path.GetFullPath(path));
+    }
+
+    /// <summary>
+    /// Resolves a logical browser name or explicit executable path to a runnable browser executable.
+    /// </summary>
+    internal static string? TryResolveExecutable(string browser)
+    {
+        if (Path.IsPathRooted(browser) && File.Exists(browser))
+        {
+            return browser;
+        }
+
+        // Probe well-known install paths before PATH lookup so logical names still work for browser app bundles or
+        // standard Windows installs that are not exposed on PATH.
+        foreach (var candidate in GetBrowserCandidates(browser))
+        {
+            if (Path.IsPathRooted(candidate))
+            {
+                if (File.Exists(candidate))
+                {
+                    return candidate;
+                }
+            }
+            else if (PathLookupHelper.FindFullPathFromPath(candidate) is { } resolvedPath)
+            {
+                return resolvedPath;
+            }
+        }
+
+        return PathLookupHelper.FindFullPathFromPath(browser);
+    }
+
+    /// <summary>
+    /// Resolves the browser's persistent user data root for shared browser-log sessions.
+    /// </summary>
+    internal static string? TryResolveUserDataDirectory(string browser, string browserExecutable)
+    {
+        var browserKind = GetBrowserKind(browser, browserExecutable);
+        if (browserKind == BrowserKind.Unknown)
+        {
+            return null;
+        }
+
+        if (OperatingSystem.IsMacOS())
+        {
+            var home = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
+            return browserKind switch
+            {
+                BrowserKind.Edge => Path.Combine(home, "Library", "Application Support", "Microsoft Edge"),
+                BrowserKind.Chrome => Path.Combine(home, "Library", "Application Support", "Google", "Chrome"),
+                _ => null
+            };
+        }
+
+        if (OperatingSystem.IsWindows())
+        {
+            var localAppData = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
+            return browserKind switch
+            {
+                BrowserKind.Edge => Path.Combine(localAppData, "Microsoft", "Edge", "User Data"),
+                BrowserKind.Chrome => Path.Combine(localAppData, "Google", "Chrome", "User Data"),
+                _ => null
+            };
+        }
+
+        var homeDirectory = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
+        return browserKind switch
+        {
+            BrowserKind.Edge => Path.Combine(homeDirectory, ".config", "microsoft-edge"),
+            // Linux Chromium packages use a different root than Google Chrome even when the requested logical browser is
+            // "chrome" but the resolved executable is chromium/chromium-browser.
+            BrowserKind.Chrome => Path.Combine(
+                homeDirectory,
+                ".config",
+                MatchesBrowser(browser, browserExecutable, "chromium", "chromium-browser") ? "chromium" : "google-chrome"),
+            _ => null
+        };
+    }
+
+    /// <summary>
+    /// Resolves a Chromium profile directory name from a directory name, profile display name, or shortcut name.
+    /// </summary>
+    internal static string ResolveProfileDirectory(string userDataDirectory, string profile)
+    {
+        ArgumentException.ThrowIfNullOrWhiteSpace(userDataDirectory);
+        ArgumentException.ThrowIfNullOrWhiteSpace(profile);
+
+        if (!Directory.Exists(userDataDirectory))
+        {
+            throw new InvalidOperationException($"Browser user data directory '{userDataDirectory}' was not found.");
+        }
+
+        if (TryResolveProfileDirectoryFromDirectoryEntries(userDataDirectory, profile) is { } directMatch)
+        {
+            return directMatch;
+        }
+
+        // Chromium stores display names in the user-data-root "Local State" file under profile.info_cache. Directory
+        // names like "Default" or "Profile 1" are stable command-line values, while display names are user-facing.
+        var localStatePath = Path.Combine(userDataDirectory, "Local State");
+        if (File.Exists(localStatePath))
+        {
+            try
+            {
+                using var localStateStream = File.OpenRead(localStatePath);
+                using var localStateDocument = JsonDocument.Parse(localStateStream);
+                if (TryResolveProfileDirectory(localStateDocument.RootElement, userDataDirectory, profile) is { } profileDirectory)
+                {
+                    return profileDirectory;
+                }
+            }
+            catch (IOException ex)
+            {
+                throw new InvalidOperationException(
+                    $"Unable to read Chromium profile metadata from '{localStatePath}' while resolving browser profile '{profile}'.",
+                    ex);
+            }
+            catch (UnauthorizedAccessException ex)
+            {
+                throw new InvalidOperationException(
+                    $"Unable to read Chromium profile metadata from '{localStatePath}' while resolving browser profile '{profile}'.",
+                    ex);
+            }
+            catch (JsonException ex)
+            {
+                throw new InvalidOperationException(
+                    $"Chromium profile metadata in '{localStatePath}' is invalid while resolving browser profile '{profile}'.",
+                    ex);
+            }
+        }
+
+        throw new InvalidOperationException(
+            $"Browser profile '{profile}' was not found under '{userDataDirectory}'. Specify the profile directory name (for example 'Default' or 'Profile 1') or a browser profile name from Chromium's profile metadata.");
+    }
+
+    /// <summary>
+    /// Resolves a profile directory from Chromium's parsed Local State metadata.
+    /// </summary>
+    internal static string? TryResolveProfileDirectory(JsonElement localStateRoot, string userDataDirectory, string profile)
+    {
+        ArgumentException.ThrowIfNullOrWhiteSpace(userDataDirectory);
+        ArgumentException.ThrowIfNullOrWhiteSpace(profile);
+
+        if (!localStateRoot.TryGetProperty("profile", out var profileElement) ||
+            !profileElement.TryGetProperty("info_cache", out var infoCacheElement) ||
+            infoCacheElement.ValueKind != JsonValueKind.Object)
+        {
+            return null;
+        }
+
+        string? match = null;
+
+        foreach (var profileEntry in infoCacheElement.EnumerateObject())
+        {
+            // Ignore stale metadata entries whose profile directories no longer exist.
+            if (!Directory.Exists(Path.Combine(userDataDirectory, profileEntry.Name)) ||
+                !MatchesBrowserProfile(profileEntry, profile))
+            {
+                continue;
+            }
+
+            // Profile display names are not unique. Force the caller to use the stable directory name when ambiguity
+            // would otherwise select an arbitrary profile.
+            if (match is not null && !string.Equals(match, profileEntry.Name, StringComparison.Ordinal))
+            {
+                throw new InvalidOperationException(
+                    $"Browser profile '{profile}' matched multiple Chromium profiles under '{userDataDirectory}'. Specify the profile directory name instead.");
+            }
+
+            match = profileEntry.Name;
+        }
+
+        return match;
+    }
+
+    /// <summary>
+    /// Gets platform-specific candidate executables for logical browser names.
+    /// </summary>
+    private static IEnumerable<string> GetBrowserCandidates(string browser)
+    {
+        if (OperatingSystem.IsMacOS())
+        {
+            return browser.ToLowerInvariant() switch
+            {
+                "msedge" or "edge" =>
+                [
+                    "/Applications/Microsoft Edge.app/Contents/MacOS/Microsoft Edge",
+                    "msedge"
+                ],
+                "chrome" or "google-chrome" =>
+                [
+                    "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome",
+                    "google-chrome",
+                    "chrome"
+                ],
+                _ => [browser]
+            };
+        }
+
+        if (OperatingSystem.IsWindows())
+        {
+            var programFiles = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles);
+            var programFilesX86 = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86);
+
+            return browser.ToLowerInvariant() switch
+            {
+                "msedge" or "edge" =>
+                [
+                    Path.Combine(programFilesX86, "Microsoft", "Edge", "Application", "msedge.exe"),
+                    Path.Combine(programFiles, "Microsoft", "Edge", "Application", "msedge.exe"),
+                    "msedge.exe"
+                ],
+                "chrome" or "google-chrome" =>
+                [
+                    Path.Combine(programFilesX86, "Google", "Chrome", "Application", "chrome.exe"),
+                    Path.Combine(programFiles, "Google", "Chrome", "Application", "chrome.exe"),
+                    "chrome.exe"
+                ],
+                _ => [browser]
+            };
+        }
+
+        return browser.ToLowerInvariant() switch
+        {
+            "msedge" or "edge" => ["microsoft-edge", "microsoft-edge-stable", "msedge"],
+            "chrome" or "google-chrome" => ["google-chrome", "google-chrome-stable", "chrome", "chromium-browser", "chromium"],
+            _ => [browser]
+        };
+    }
+
+    private static string? TryResolveProfileDirectoryFromDirectoryEntries(string userDataDirectory, string profile)
+    {
+        foreach (var directoryPath in Directory.EnumerateDirectories(userDataDirectory))
+        {
+            var directoryName = Path.GetFileName(directoryPath);
+            if (string.Equals(directoryName, profile, StringComparison.OrdinalIgnoreCase))
+            {
+                return directoryName;
+            }
+        }
+
+        return null;
+    }
+
+    /// <summary>
+    /// Classifies a browser from both the requested value and the resolved executable path.
+    /// </summary>
+    private static BrowserKind GetBrowserKind(string browser, string browserExecutable)
+    {
+        if (MatchesBrowser(browser, browserExecutable, "msedge", "edge", "microsoft-edge"))
+        {
+            return BrowserKind.Edge;
+        }
+
+        if (MatchesBrowser(browser, browserExecutable, "chrome", "google-chrome", "chromium", "chromium-browser"))
+        {
+            return BrowserKind.Chrome;
+        }
+
+        return BrowserKind.Unknown;
+    }
+
+    private static bool MatchesBrowser(string browser, string browserExecutable, params string[] names)
+    {
+        var browserLower = browser.ToLowerInvariant();
+        var executableLower = browserExecutable.ToLowerInvariant();
+
+        foreach (var name in names)
+        {
+            if (browserLower == name ||
+                Path.GetFileNameWithoutExtension(browserLower) == name ||
+                executableLower.Contains(name, StringComparison.Ordinal))
+            {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    private static bool MatchesBrowserProfile(JsonProperty profileEntry, string profile)
+    {
+        return string.Equals(profileEntry.Name, profile, StringComparison.OrdinalIgnoreCase) ||
+            MatchesBrowserProfileProperty(profileEntry.Value, "name", profile) ||
+            MatchesBrowserProfileProperty(profileEntry.Value, "shortcut_name", profile);
+    }
+
+    private static bool MatchesBrowserProfileProperty(JsonElement profileElement, string propertyName, string profile)
+    {
+        return profileElement.TryGetProperty(propertyName, out var propertyElement) &&
+            propertyElement.ValueKind == JsonValueKind.String &&
+            string.Equals(propertyElement.GetString(), profile, StringComparison.OrdinalIgnoreCase);
+    }
+
+    private enum BrowserKind
+    {
+        Unknown,
+        Edge,
+        Chrome
+    }
+}
